fix playback crash; motionlayout elevation
