--- app/src/main/res/drawable/divider.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle" >
-
-    <size
-        android:height="2dp"/>
-
-    <solid android:color="@android:color/white" />
-
-</shape>
