+++ app/src/main/res/layout/fragment_broadcast_details.xml
@@ -39,6 +39,7 @@
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:background="@color/colorPrimary50a"
+            android:elevation="@dimen/spacing_medium"
             app:layout_constraintBottom_toBottomOf="@id/broadcast_date"
             app:layout_constraintTop_toTopOf="@+id/broadcast_parent_show"
             app:layout_constraintVertical_bias="1.0"
@@ -54,6 +55,7 @@
             android:text="@{vm.showLiveData.name}"
             android:paddingStart="@dimen/spacing_large"
             android:paddingTop="@dimen/spacing_small"
+            android:elevation="@dimen/spacing_medium"
             app:layout_constrainedWidth="true"
             app:layout_constraintBottom_toTopOf="@id/broadcast_date"
             app:layout_constraintEnd_toStartOf="@id/downloadDeleteIcon"
@@ -71,6 +73,7 @@
             android:paddingStart="@dimen/spacing_large"
             android:paddingBottom="@dimen/spacing_small"
             android:maxLines="1"
+            android:elevation="@dimen/spacing_medium"
             app:layout_constraintBottom_toBottomOf="@id/broadcast_detail_image"
             app:layout_constraintStart_toStartOf="parent"
             tools:text="Jan 01 2019" />
@@ -86,6 +89,7 @@
             android:layout_marginTop="@dimen/spacing_medium"
             android:layout_marginBottom="@dimen/spacing_medium"
             android:layout_marginEnd="@dimen/spacing_large"
+            android:elevation="@dimen/spacing_large"
             android:src="@drawable/exo_controls_play"
             android:visibility="gone"
             tools:visibility="visible" />
@@ -101,6 +105,7 @@
             app:layout_constraintEnd_toStartOf="@id/archivePlayButton"
             android:layout_marginEnd="@dimen/spacing_large"
             android:visibility="gone"
+            android:elevation="@dimen/spacing_large"
             tools:visibility="visible"
             android:tag="download"/>
 
@@ -133,15 +138,6 @@
 
         </LinearLayout>
 
-        <ImageView
-            android:id="@+id/broadcast_divider"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:src="@drawable/divider"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/description_container"/>
-
         <TextView
             android:id="@+id/noTracksMessage"
             android:layout_width="wrap_content"
@@ -149,7 +145,7 @@
             android:text="@string/no_tracks"
             style="@style/TextAppearance.MaterialComponents.Subtitle1"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/broadcast_divider"
+            app:layout_constraintTop_toBottomOf="@id/description_container"
             android:layout_marginTop="@dimen/spacing_large"
             android:layout_marginStart="@dimen/spacing_large"
             android:layout_marginEnd="@dimen/spacing_large"
@@ -166,7 +162,7 @@
             android:paddingBottom="@dimen/spacing_medium"
             android:clipToPadding="false"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/broadcast_divider"
+            app:layout_constraintTop_toBottomOf="@id/description_container"
             tools:visibility="gone"/>
 
     </androidx.constraintlayout.motion.widget.MotionLayout>
