+++ app/src/main/res/xml/scene_broadcast_details.xml
@@ -79,6 +79,17 @@
             app:layout_constraintTop_toBottomOf="@id/broadcast_detail_image"
             app:visibilityMode="ignore"/>
 
+        <Constraint
+            android:id="@+id/noTracksMessage"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/description_container"
+            android:layout_marginTop="@dimen/spacing_large"
+            android:layout_marginStart="@dimen/spacing_large"
+            android:layout_marginEnd="@dimen/spacing_large"
+            app:visibilityMode="ignore"/>
+
     </ConstraintSet>
 
     <ConstraintSet android:id="@+id/collapsed">
@@ -146,6 +157,17 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/broadcast_detail_image"
+            app:visibilityMode="ignore"/>
+
+        <Constraint
+            android:id="@+id/noTracksMessage"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/description_container"
+            android:layout_marginTop="@dimen/spacing_large"
+            android:layout_marginStart="@dimen/spacing_large"
+            android:layout_marginEnd="@dimen/spacing_large"
             app:visibilityMode="ignore"/>
     </ConstraintSet>
